pyproject.toml
[tool.poetry]
name = "generalbot"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.8"
py-cord = "^1.7.3"
SQLAlchemy = "^1.4.23"
alembic = "^1.7.1"
sanic = "^21.6.2"
typing-extensions = "^3.10.0"
youtube_dl = "^2021.6.6"
captcha = "^0.3"
orjson = "^3.6.3"
discord-debug = "^1.0.1"
jishaku = "^2.3.0"
wheel = "^0.37.0"
mongoengine = "^0.23.1"
Flask = "^2.0.1"
psutil = "^5.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
